Convert AuthenticationHelpers module to a patch
This way, we don't need to load ApplicationController in a
config.to_prepare block either.
